:art: ResultCode 扩展

+++ lingting-core/src/main/kotlin/live/lingting/framework/api/ApiResultCode.kt
@@ -4,7 +4,7 @@ package live.lingting.framework.api
  * @author lingting 2022/9/19 13:56
  */
 enum class ApiResultCode(
-    override val code: Long,
+    override val code: Int,
     override val message: String,
 ) : ResultCode {
     /**
